code/data_utils/ml_studio_FasterRCNN_backend/requirements.txt
Flask~=2.3
colorama~=0.4
requests~=2.31
semver~=3.0.2
pillow~=10.3
label_studio_tools @ git+https://github.com/HumanSignal/label-studio-tools.git@master
# label_studio_sdk from git master branch
label_studio_sdk @ git+https://github.com/heartexlabs/label-studio-sdk.git@master
torch==2.3.0
torchvision==0.18.0
requests==2.32.2
python-dotenv~=1.0.1